# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: ResearchFlow API
  description: |
    ResearchFlow API - Orchestrator Service

    A comprehensive research workflow management platform with:
    - User authentication and authorization
    - Paper library and annotation management
    - Organization and project management
    - AI-powered research insights and analysis
    - Data extraction and processing
    - Version control and collaboration features

    All requests should include the Authorization header with a JWT Bearer token,
    except for public endpoints like registration, login, and authentication status.
  version: 1.0.0
  contact:
    name: ResearchFlow Team
    url: https://researchflow.io
  license:
    name: Proprietary
    url: https://researchflow.io/license
servers:
  - url: https://api.researchflow.io/api
    description: Production environment
  - url: http://localhost:5000/api
    description: Local development environment
  - url: http://orchestrator:5000/api
    description: Docker internal environment
security:
  - BearerAuth: []
  - cookieAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token in Bearer format
    cookieAuth:
      type: apiKey
      in: cookie
      name: refreshToken
      description: HTTP-only cookie containing refresh token
  schemas:
...
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Error message
        details:
          oneOf:
            - type: object
              description: Additional error details
            - type: array
              description: Array of validation errors
        code:
          type: string
          description: Error code for programmatic handling
        request_id:
          type: string
          description: Request identifier for tracking
      example:
        error: "Validation failed"
        details:
          - field: "email"
            message: "Invalid email format"
        code: "VALIDATION_ERROR"
...
    User:
      type: object
      required:
        - id
        - email
        - displayName
        - role
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
          description: User unique identifier
        email:
          type: string
          format: email
          description: User email address
        firstName:
          type: string
          description: First name
        lastName:
          type: string
          description: Last name
        displayName:
          type: string
          description: Display name for profile
        profileImageUrl:
          type: string
          format: uri
          description: URL to profile image
        role:
          type: string
          enum: [ADMIN, RESEARCHER, VIEWER]
          description: User role in system
        orgId:
          type: string
          format: uuid
          description: Organization ID
        createdAt:
          type: string
          format: date-time
          description: Account creation timestamp
        updatedAt:
          type: string
          format: date-time
          description: Last update timestamp
    RegisterRequest:
      type: object
      required:
        - email
        - password
        - displayName
      properties:
        email:
          type: string
          format: email
          description: User email address
        password:
          type: string
          format: password
          minLength: 8
          description: Password (minimum 8 characters)
        displayName:
          type: string
          description: Display name
        firstName:
          type: string
          description: First name (optional)
        lastName:
          type: string
          description: Last name (optional)
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          description: User email address
        password:
          type: string
          format: password
          description: User password
    AuthResponse:
      type: object
      required:
        - user
        - accessToken
      properties:
        message:
          type: string
        user:
          $ref: '#/components/schemas/User'
        accessToken:
          type: string
          format: jwt
          description: JWT access token (valid for 1 hour)
    PasswordResetRequest:
      type: object
      required:
        - token
        - newPassword
      properties:
        token:
          type: string
          description: Password reset token from email
        newPassword:
          type: string
          format: password
          minLength: 8
          description: New password

    # ===========================================================================
    # Organization Schemas
    # ===========================================================================
...
  responses:
    UnauthorizedError:
      description: Authentication failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Unauthorized"
            code: "UNAUTHORIZED"
...
    ValidationError:
      description: Validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Internal server error"
            code: "INTERNAL_ERROR"

# ===========================================================================
# Paths / Endpoints
# ===========================================================================
paths:
...
  /auth/register:
    post:
      tags:
        - Authentication
      summary: Register a new user account
      operationId: registerUser
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Registration successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
              example:
                message: "Registration successful"
                user:
                  id: "550e8400-e29b-41d4-a716-446655440000"
                  email: "user@example.com"
                  displayName: "John Doe"
                  role: "RESEARCHER"
                accessToken: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'
  /auth/login:
    post:
      tags:
        - Authentication
      summary: Login with email and password
      operationId: loginUser
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'
  /auth/refresh:
    post:
      tags:
        - Authentication
      summary: Refresh access token using refresh token
      operationId: refreshToken
      security:
        - cookieAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken:
                  type: string
                  description: Refresh token (can also use cookie)
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  accessToken:
                    type: string
                    format: jwt
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalError'
  /auth/logout:
    post:
      tags:
        - Authentication
      summary: Logout and revoke tokens
      operationId: logoutUser
      responses:
        '200':
          description: Logged out successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Logged out successfully"
        '500':
          $ref: '#/components/responses/InternalError'
  /auth/logout-all:
    post:
      tags:
        - Authentication
      summary: Logout from all devices
      operationId: logoutAllDevices
      responses:
        '200':
          description: Logged out from all devices
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Logged out from all devices"
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalError'
  /auth/user:
    get:
      tags:
        - Authentication
      summary: Get current authenticated user
      operationId: getCurrentUser
      responses:
        '200':
          description: Current user data
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /auth/forgot-password:
    post:
      tags:
        - Authentication
      summary: Request a password reset link
      operationId: forgotPassword
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
              properties:
                email:
                  type: string
                  format: email
      responses:
        '200':
          description: Password reset link sent (if user exists)
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'
  /auth/reset-password:
    post:
      tags:
        - Authentication
      summary: Reset password with reset token
      operationId: resetPassword
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordResetRequest'
      responses:
        '200':
          description: Password reset successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'
  /auth/change-password:
    post:
      tags:
        - Authentication
      summary: Change password (requires authentication)
      operationId: changePassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - currentPassword
                - newPassword
              properties:
                currentPassword:
                  type: string
                  format: password
                newPassword:
                  type: string
                  format: password
                  minLength: 8
      responses:
        '200':
          description: Password changed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalError'

  # ===========================================================================
  # Organizations Endpoints (/api/org/*)
  # ===========================================================================
...